requirements.txt
gradio>=4.0.0
huggingface-hub>=0.19.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
requests>=2.31.0